update: restructure components and improve DSL consistency in Router and Sample modules
