+++ renlin-router/src/commonMain/kotlin/net/kigawa/renlin/router/link/Link.kt
@@ -1,7 +1,6 @@
 package net.kigawa.renlin.router.link
 
-import net.kigawa.renlin.tag.Tag
+import net.kigawa.renlin.component.component
 import net.kigawa.renlin.tag.a
-import net.kigawa.renlin.tag.component.Component1
 
-val link = a.component {  }
+val link = a.component { }
